pyproject.toml
[project]
name = "rag-notion"
version = "0.1.0"
description = "RAG pipeline Notion — Offline ingestion + Online API (LangChain, Qdrant, Cohere, Mistral)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "langchain>=0.3",
    "langchain-core>=0.3",
    "langchain-cohere>=0.2",
    "langchain-mistralai>=0.1",
    "langchain-qdrant>=0.2",
    "langchain-community>=0.3",
    "qdrant-client>=1.7",
    "notion-client>=2.2",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
